+++ .circleci/tests/default_values_charts_test.go
@@ -1,0 +1,28 @@
+package tests
+
+import (
+	"path"
+	"path/filepath"
+)
+
+func (s *helmRepoSuite) TestChartTemplateWithDefaultValues() {
+	charts := s.getChartsInDirectory(chartsRootDir)
+
+	for _, chartName := range charts {
+		s.Run(chartName, func() {
+			chartDir := s.getChartDirectory(chartName)
+			chartDirFullPath, err := filepath.Abs(chartDir)
+			s.NoError(err)
+			defaultValuesFilePath := path.Join(chartDir, "values.yaml")
+
+			runGoldenHelmTest(s.T(), &goldenHelmTest{
+				ChartPath:          chartDirFullPath,
+				Release:            "release-test",
+				Namespace:          "release-helm-namespace",
+				GoldenFileName:     chartName,
+				ValuesFile:         defaultValuesFilePath,
+				GoldenSubDirectory: "default",
+			})
+		})
+	}
+}
